+++ src/main/java/pl/pjatk/store/repository/ArticleRepository.java
@@ -5,4 +5,4 @@ import org.springframework.stereotype.Repository;
 import pl.pjatk.store.model.Article;
 
 @Repository
-public interface ArticleRepository extends JpaRepository<Article, Long> { }
+public interface ArticleRepository extends JpaRepository<Article, Integer> { }
